Automatically adding the user to the group video
Using Ubuntu 22.04 the user running autodarts needs to be in the group video to be able to access the cameras. Therefore I recommend to add mentioned lines. 

In case the group is not available or not needed this command usually does not harm the system BUT it grants rights which cannot be safely undone.

Please be aware that Ubuntu, Debian, Suse, Mandriva and others to have an operational packaging system which handles these things much saver than the install script here. Nevertheless it might help others to avoid mistakes.